+++ src/lib/components/tools/Notes.svelte
